Fixing minor FindBug hits and removing test code.
git-svn-id: file:///home/dweiss/carrot2/migration/carrot2.svn/trunk@4935 7ff1d41c-760d-0410-a7ff-a3a56f310b35

+++ core/carrot2-util-text/src-test/org/carrot2/text/linguistic/DefaultLexicalDataFactoryTest.java
@@ -247,10 +247,16 @@ public class DefaultLexicalDataFactoryTest
         try
         {
             File tempFile = File.createTempFile("tempdir", null);
-            tempFile.delete();
-            if (!tempFile.mkdir())
+            if (tempFile.exists() && !tempFile.delete()) 
+            {
+                throw new RuntimeException("Could not delete temp. folder: " 
+                    + tempFile.getAbsolutePath());
+            }
+            if (!tempFile.mkdir()) 
+            {
                 throw new RuntimeException("Could not create temp. folder: " 
                     + tempFile.getAbsolutePath());
+            }
 
             tempFile.deleteOnExit();
